Add ProxyResource class
