+++ sdk/core/azure-core-management/src/main/java/com/azure/core/management/ProxyResource.java
@@ -1,0 +1,56 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+package com.azure.core.management;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+/**
+ * The Proxy Resource model.
+ */
+public class ProxyResource {
+    /**
+     * Resource Id.
+     */
+    @JsonProperty(access = JsonProperty.Access.WRITE_ONLY)
+    private String id;
+
+    /**
+     * Resource name.
+     */
+    @JsonProperty(access = JsonProperty.Access.WRITE_ONLY)
+    private String name;
+
+    /**
+     * Resource type.
+     */
+    @JsonProperty(access = JsonProperty.Access.WRITE_ONLY)
+    private String type;
+
+    /**
+     * Get the id value.
+     *
+     * @return the id value
+     */
+    public String getId() {
+        return this.id;
+    }
+
+    /**
+     * Get the name value.
+     *
+     * @return the name value
+     */
+    public String getName() {
+        return this.name;
+    }
+
+    /**
+     * Get the type value.
+     *
+     * @return the type value
+     */
+    public String getType() {
+        return this.type;
+    }
+}
