+++ sdk/core/azure-core-management/src/main/java/com/azure/core/management/Resource.java
@@ -10,25 +10,7 @@ import java.util.Map;
 /**
  * The Resource model.
  */
-public class Resource {
-    /**
-     * Resource Id.
-     */
-    @JsonProperty(access = JsonProperty.Access.WRITE_ONLY)
-    private String id;
-
-    /**
-     * Resource name.
-     */
-    @JsonProperty(access = JsonProperty.Access.WRITE_ONLY)
-    private String name;
-
-    /**
-     * Resource type.
-     */
-    @JsonProperty(access = JsonProperty.Access.WRITE_ONLY)
-    private String type;
-
+public class Resource extends ProxyResource {
     /**
      * Resource location.
      */
@@ -39,33 +21,6 @@ public class Resource {
      * Resource tags.
      */
     private Map<String, String> tags;
-
-    /**
-     * Get the id value.
-     *
-     * @return the id value
-     */
-    public String getId() {
-        return this.id;
-    }
-
-    /**
-     * Get the name value.
-     *
-     * @return the name value
-     */
-    public String getName() {
-        return this.name;
-    }
-
-    /**
-     * Get the type value.
-     *
-     * @return the type value
-     */
-    public String getType() {
-        return this.type;
-    }
 
     /**
      * Get the location value.
